Unify the Text model and remove subclassing
This involves removing the baseclass model BaseText and merging the
fields from the BaseText model to the new model Text, and removing the
model AuthorityText. This also involves quite some data migrating. Since
this change introduces fundamental changes in the database structure, it
is not compatible with previous DB models. It is not possible to jump
forth and back between the two structures in the development tree
easily.

This commit introduces the following changes:

Squash commits in Commentaries and Manuscripts

Remove Commentary model

Rename Text.commentator to Text.author

Update CommentatorAlternative to AuthorAlternative

With consequences in admin imports and calls
Update __str__ method on AuthorAlternative

Add text_type field to Text model

This prepares for the deletion of AuthorityText and BaseText models

Let Text.commentary_on refer to Text itself

With a limitation to authority texts

Remove AuthorityText and BaseText models

Add extra text type choices

'Summa or tractatus' and 'Compilation or excerpts'